Add new stroybook
