+++ .storybook/addons.js
@@ -1,1 +1,1 @@
-// import 'storybook-addon-jsx/register';
+import '@storybook/addon-actions/register';
