+++ theme.css
@@ -62,7 +62,7 @@ body {
 .card {
   box-shadow: rgba(0, 0, 0, 0.117647) 0 1px 6px, rgba(0, 0, 0, 0.117647) 0 1px 4px;
   width: 100%;
-  max-width: 400px;
+  max-width: 600px;
   margin: 10px;
   background-color: #fff;
 }
